finalizeing the workspace structure dinally fixed the workspace. though i made the workapce so that it is a package itself maybe it destroies the poupous of the tutorial

+++ add_one/src/lib.rs
@@ -1,4 +1,4 @@
-pub fn add_two(number: i32) -> u64 {
+pub fn add_two(number: i32) -> i32 {
     number + 2
 }
