+++ add_two/src/lib.rs
@@ -1,4 +1,4 @@
-pub fn add_one(number: i32) -> u64 {
+pub fn add_one(number: i32) -> i32 {
     number + 1
 }
